Fix factory() support. Convert Pimple usage from delegation to trait.
When adding a unit-test for `factory()`, it became clear that the
`factory()` wrapper wasn't working properly.

After some debugging, I realized the functionality would be easier/more
consistent with 2-3 lines of changes in `\Pimple\Container`.  Alas,
extending the class doesn't fly b/c `\Pimple\Container` is protective of the
inernals.  It's understandble -- since extending the class and changing this
behavior would produce a container class that's not genuinely
interchangeable.

Since the goal here is more to use the same implementation+style but to
change some API elements, it seems more accurate to model this as a
`\Pimple\ContainerTrait` used by `\Clippy\Container` consumer.